requirements.txt
crawl4ai>=0.2.77
playwright>=1.40.0
asyncio
aiohttp
beautifulsoup4
lxml
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tweepy>=4.14.0
facebook-sdk>=3.1.0
schedule>=1.2.0
python-dateutil>=2.8.2
flask>=2.3.0